// frontendclient/src/app/api/generated/education/education.service.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import {
  HttpClient
} from '@angular/common/http';

import {
  Injectable,
  inject
} from '@angular/core';

import type {
  DeepNonNullable
} from '@orval/core';

import type {
  Education,
  FindByFromYearParams,
  FindByLocationParams,
  SearchByNameParams
} from '../../models';

import { customHttpClient } from '../../custom-http-client';






@Injectable({ providedIn: 'root' })
export class EducationService {
  private readonly http = inject(HttpClient);
/**
 * Retrieve a specific education entry by its ID
 * @summary Get education by ID
 */
 getEducationById<TData = Education>(
    id: string,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education/${id}`, method: 'GET'
    },
      this.http,
      );
    }
  /**
 * Update an existing education entry
 * @summary Update education record
 */
 updateEducation<TData = Education>(
    id: string,
    education: Education,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education/${id}`, method: 'PUT',
      headers: {'Content-Type': 'application/json', },
      data: education
    },
      this.http,
      );
    }
  /**
 * Remove an education entry
 * @summary Delete education record
 */
 deleteEducation<TData = void>(
    id: string,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education/${id}`, method: 'DELETE'
    },
      this.http,
      );
    }
  /**
 * Retrieve a list of all education entries
 * @summary Get all education records
 */
 getAllEducation<TData = Education[]>(
    
 ) {
      return customHttpClient<TData>(
      {url: `/api/education`, method: 'GET'
    },
      this.http,
      );
    }
  /**
 * Add a new education entry
 * @summary Create education record
 */
 createEducation<TData = Education>(
    education: Education,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education`, method: 'POST',
      headers: {'Content-Type': 'application/json', },
      data: education
    },
      this.http,
      );
    }
  /**
 * Search for education entries by from year
 * @summary Find education by starting year
 */
 findByFromYear<TData = Education[]>(
    params: DeepNonNullable<FindByFromYearParams>,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education/search/year`, method: 'GET',
        params
    },
      this.http,
      );
    }
  /**
 * Search for education entries by institution name
 * @summary Find education by institution name
 */
 searchByName<TData = Education[]>(
    params: DeepNonNullable<SearchByNameParams>,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education/search/name`, method: 'GET',
        params
    },
      this.http,
      );
    }
  /**
 * Search for education entries by location
 * @summary Find education by location
 */
 findByLocation<TData = Education[]>(
    params: DeepNonNullable<FindByLocationParams>,
 ) {
      return customHttpClient<TData>(
      {url: `/api/education/search/location`, method: 'GET',
        params
    },
      this.http,
      );
    }
  };

export type GetEducationByIdClientResult = NonNullable<Education>
export type UpdateEducationClientResult = NonNullable<Education>
export type DeleteEducationClientResult = NonNullable<void>
export type GetAllEducationClientResult = NonNullable<Education[]>
export type CreateEducationClientResult = NonNullable<Education>
export type FindByFromYearClientResult = NonNullable<Education[]>
export type SearchByNameClientResult = NonNullable<Education[]>
export type FindByLocationClientResult = NonNullable<Education[]>