// frontendclient/src/app/api/models/education.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface Education {
  id?: string;
  name?: string;
  instituitionName?: string;
  fromYear?: string;
  toYear?: string;
  location?: string;
  cgpa?: string;
  lastUpdated?: string;
  isVisible?: boolean;
}